+++ src/index.ts
@@ -1,0 +1,41 @@
+import ConnectedLoadingBar from './loading_bar.tsx';
+import { LoadingBar } from './loading_bar.tsx';
+import type { LoadingBarContainerProps, LoadingBarProps } from './loading_bar.tsx';
+import loadingBarMiddleware from './loading_bar_middleware.ts';
+import type { MiddlewareConfig } from './loading_bar_middleware.ts';
+import {
+    DEFAULT_SCOPE,
+    HIDE,
+    hideLoading,
+    loadingBarReducer,
+    RESET,
+    resetLoading,
+    SHOW,
+    showLoading,
+} from './loading_bar_ducks.ts';
+import type { LoadingBarAction, LoadingBarState } from './loading_bar_ducks.ts';
+import ImmutableLoadingBar from './immutable.tsx';
+
+export {
+    DEFAULT_SCOPE,
+    HIDE,
+    hideLoading,
+    ImmutableLoadingBar,
+    LoadingBar,
+    loadingBarMiddleware,
+    loadingBarReducer,
+    RESET,
+    resetLoading,
+    SHOW,
+    showLoading,
+};
+
+export type {
+    LoadingBarAction,
+    LoadingBarContainerProps,
+    LoadingBarProps,
+    LoadingBarState,
+    MiddlewareConfig,
+};
+
+export default ConnectedLoadingBar;